+++ packages/vue/src/plugin.ts
@@ -1,7 +1,7 @@
 import Vue, { PluginObject } from "vue";
 
 import Link from "./Link";
-
+import AsyncLink from "./AsyncLink";
 import focus from "./focus";
 
 import { CuriRouter } from "@curi/types";
@@ -14,6 +14,7 @@ export interface CuriPluginOptions {
 const CuriPlugin: PluginObject<CuriPluginOptions> = {
   install: function(_Vue: typeof Vue, options: CuriPluginOptions) {
     _Vue.component(Link.name, Link);
+    _Vue.component(AsyncLink.name, AsyncLink);
 
     _Vue.directive("curi-focus", focus);