+++ packages/vue/types/utils/canNavigate.d.ts
@@ -1,0 +1,1 @@
+export declare function canNavigate(event: MouseEvent): boolean;